requirements.txt
Django==1.6.2
PyMySQL==0.6.1
django-allauth==0.16.1
djangorestframework==2.3.13
oauthlib==0.6.1
python3-openid==3.0.3
requests==2.2.1
requests-oauthlib==0.4.0

// Remove?
redis==2.9.1
django-constance==0.6